Move string relation functions from options_helper.* to string_util
This is part of an effort to club all string related utility functions
into a common place, in string_util.
options_helper.[h,cc] has many string utility functions that could be
useful in other places, and need to be in a more common place like
string_util.

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Blame Revision:
