+++ options/options_test.cc
@@ -25,6 +25,7 @@
 #include "rocksdb/utilities/leveldb_options.h"
 #include "util/random.h"
 #include "util/stderr_logger.h"
+#include "util/string_util.h"
 #include "util/testharness.h"
 #include "util/testutil.h"
